requirements.txt
streamlit
pymupdf
python-docx
google-genai
python-dotenv